+++ src/lib/utils/types.ts
@@ -1,4 +1,0 @@
-
-export type Maybe<T> = T | undefined;
-
-export type Primitives = string | number | boolean;